requirements.txt
absl-py==2.4.0
altair==6.0.0
annotated-types==0.7.0
arcosparse==0.4.2
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
boto3==1.42.44
botocore==1.42.44
cachetools==6.2.6
Cartopy==0.25.0
certifi==2026.1.4
cffi==2.0.0
cftime==1.6.5
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
copernicusmarine==2.3.0
cryptography==46.0.5
cycler==0.12.1
dask==2026.1.2
donfig==0.8.1.post1
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-crc32c==1.8.0
google-pasta==0.2.0
grpcio==1.78.1
h5netcdf==1.8.1
h5py==3.15.1
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
locket==1.0.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.1.0
narwhals==2.16.0
netCDF4==1.7.4
numcodecs==0.16.5
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
partd==1.4.2
pillow==12.1.0
protobuf==4.25.8
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
pyproj==3.7.2
pyshp==3.0.3
pystac==1.14.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
scipy==1.17.0
seaborn==0.13.2
semver==3.0.4
shapely==2.1.2
six==1.17.0
smmap==5.0.2
streamlit==1.54.0
tenacity==9.1.4
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.6
wrapt==1.14.2
xarray==2026.1.0
zarr==3.1.5
zipp==3.23.0
